move agentic test.Models to test module
