+++ langchain4j-agentic-a2a/pom.xml
@@ -50,14 +50,8 @@
 
         <dependency>
             <groupId>dev.langchain4j</groupId>
-            <artifactId>langchain4j-ollama</artifactId>
-            <version>1.9.0-SNAPSHOT</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>dev.langchain4j</groupId>
-            <artifactId>langchain4j-open-ai</artifactId>
-            <version>1.9.0-SNAPSHOT</version>
+            <artifactId>langchain4j-test</artifactId>
+            <version>1.9.0-beta16-SNAPSHOT</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
