+++ langchain4j-agentic-patterns/src/test/java/dev/langchain4j/agentic/patterns/p2p/researcher/P2PResearcherIT.java
@@ -4,7 +4,7 @@ import dev.langchain4j.agentic.AgenticServices;
 import dev.langchain4j.agentic.patterns.p2p.P2PPlanner;
 import org.junit.jupiter.api.Test;
 
-import static dev.langchain4j.agentic.patterns.Models.baseModel;
+import static dev.langchain4j.test.model.Models.baseModel;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class P2PResearcherIT {
