+++ langchain4j-agentic-patterns/src/test/java/dev/langchain4j/agentic/patterns/p2p/researcher/SupervisorResearcher.java
@@ -1,7 +1,7 @@
 package dev.langchain4j.agentic.patterns.p2p.researcher;
 
-import static dev.langchain4j.agentic.patterns.Models.baseModel;
-import static dev.langchain4j.agentic.patterns.Models.plannerModel;
+import static dev.langchain4j.test.model.Models.baseModel;
+import static dev.langchain4j.test.model.Models.plannerModel;
 
 import dev.langchain4j.agentic.AgenticServices;
 import dev.langchain4j.agentic.supervisor.SupervisorAgent;
