+++ langchain4j-test/pom.xml
@@ -40,6 +40,18 @@
             <version>1.9.0-SNAPSHOT</version>
         </dependency>
 
+        <dependency>
+            <groupId>dev.langchain4j</groupId>
+            <artifactId>langchain4j-ollama</artifactId>
+            <version>1.9.0-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>dev.langchain4j</groupId>
+            <artifactId>langchain4j-open-ai</artifactId>
+            <version>1.9.0-SNAPSHOT</version>
+        </dependency>
+
         <!-- test dependencies -->
 
         <dependency>
